# apis/backend-api/web-framework/openapi.json
{
  "components": {
    "securitySchemes": {
      "sdkworkAccessToken": {
        "description": "Signed JWT access_token (`header.payload.signature`). Required on all non-open-api routes, including public and refresh-token entrypoints, for tenant isolation.",
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "sdkworkDualToken": {
        "description": "Protected routes require Authorization: Bearer <auth_token JWT> and Access-Token: <access_token JWT>.",
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "sdkworkAgentToken": {
        "description": "Backend agent bootstrap token for /backend/v3/api/agent/* routes. Resolves via api-key auth-mode without dual-token JWT.",
        "in": "header",
        "name": "X-SDKWork-Agent-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "SDKWork Web Framework Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/backend/v3/api/web-framework/cors-policies": {
      "get": {
        "operationId": "webFramework.corsPolicies.list",
        "parameters": [
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "sdkworkDualToken": []
          }
        ],
        "summary": "webFramework.corsPolicies.list",
        "tags": [
          "WebFramework"
        ],
        "x-sdkwork-alternate-permissions": [
          "web-framework.platform.read"
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web-framework.tenant.admin",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-required-surface": "organizationMember",
        "x-sdkwork-route-auth": "dual-token"
      },
      "put": {
        "operationId": "webFramework.corsPolicies.upsert",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Payload Too Large"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "sdkworkDualToken": []
          }
        ],
        "summary": "webFramework.corsPolicies.upsert",
        "tags": [
          "WebFramework"
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web-framework.tenant.admin",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-required-surface": "organizationMember",
        "x-sdkwork-route-auth": "dual-token"
      }
    },
...
  }
}